Dispaly last editor only
